refactor : 카테고리 코드 분리
